requirements.txt
pynput==1.7.6
python-dotenv==1.0.0
six==1.16.0
